+++ RELEASE-NOTES.md
@@ -43,3 +43,6 @@
 DAO Info Page
 - ecosystemTvl added
 - circulating INK added NOTE: Calculates based on harvested INK. IE, unharvested INK is not circulating based on this calculation.
+- daily transaction count added
+- prices added
+- DAO holdings added, including multisig and farm addresses